backend/requirements.txt
fastapi
uvicorn[standard]
groq
python-dotenv
httpx
supabase
pydantic
python-multipart
jose
passlib[bcrypt]
email-validator
